requirements.txt
fastapi>=0.110.0
uvicorn>=0.23.0
pydantic>=2.6.0
pydantic-settings>=2.2.1
PyYAML>=6.0.1
PyJWT>=2.8.0
pytest>=8.0.0
httpx>=0.26.0
huggingface_hub>=0.20.0
SQLAlchemy>=2.0.0
psycopg[binary]>=3.1.0

# Image generation (optional - for local image models)
torch>=2.0.0
diffusers>=0.25.0
transformers>=4.35.0
accelerate>=0.25.0

# 3D preview rendering (optional - for static mesh previews)
trimesh>=4.3.0
pyrender>=0.1.45
Pillow>=10.0.0
numpy>=1.26.0
PyOpenGL>=3.1.7
pyglet>=2.0.10
